// core/types.ts
export interface PolicyConfig {
  version: "1";
  maxValueWei: string;
  maxApprovalAmount: string;
  contractAllowlist: string[];
  tokenAllowlist: string[];
  allowedChains: number[];
  recipientAllowlist: string[];
  maxRiskScore: number;
  requireApprovalAbove: { valueWei: string };
  maxTxPerHour: number;
}